requirements.txt
easydict
hydra-core
gdown
huggingface-hub
omegaconf
pandas
Pillow
protobuf
PyYAML
scipy
seaborn
simcse
tqdm
timm
torch
torchvision
wandb
yacs
rich
ftfy